Split integration.
